Rename classes to camel case to make PyCharm happy
